fix(Memento): Return all properties with correct casing in memento
All properties defined on the model should now be returned when calling `getMemento`.
The casing used is the casing defined as the property name instead of the column.
If no value has been set, like in the case of a new entity, the default value of "" (empty string) will be used.