feat(BaseEntity): Improve null handling
Null values are retrieved as empty strings from the database.
By default, empty strings are serialized back to nulls when saving to the database.
This feature can be turned off per property by setting the `convertToNull` attribute to `false`.
Additionally, the `nullValue` property can be set to a custom value.  When this is the case-sensitive
value of the property when saving, `null` will be inserted into the database.