refactor(QuickCollection): Extract getMemento from $renderData
QuickCollection had included a helper function `$renderData`
that would call `getMemento()` on all internal objects and return it
as an array.  Since the `$renderData` name is used for ColdBox
automatic data marshalling, we are extracting the logic to a
`getMemento` function for more semantic reuse.  The `$renderData`
method now calls `getMemento`.