fix(Relationships): Fix incorrect and slow relationships
The original implementation of relationships had some naive
implementations that were both slow and potentially incorrect.
Some of this was hidden by simplistic test cases which have been
updated.  This commit fixes those issues and brings the API
more in line with Eloquent from Laravel. (This library is _**heavily**_
inspired by Eloquent.)

BREAKING CHANGE: The function signature for all the relationship
methods have changed.  It is possible to not need any changes if
you relied on the conventions.  If you did not you may need to change
some argument names or argument order to get the relationships
working again.