feat(Scopes): Enable named and default arguments in scopes
Extra arguments passed to scope are passed in order to the
scope defined on the entity.  Default arguments can also be set
now for individual scope arguments making it more close to a
normal function signature.

BREAKING CHANGE: Extra arguments to a scope function used
to be passed as a single struct called `args`.  They are now passed
individually in the order they were passed to the scope.