perf(BaseEntity): Use shallow duplicate where possible
Lucee allows for shallow `duplicate` which is insanely fast.
ACF doesn't have this capability and it fails to run if it
is passed even in an `if` statement.  So we split the entity
creation off in to separate components so Lucee can be
super fast and ACF can stop complaining.