feat(BaseService): Create a BaseService for virtual entity services
Akin to cborm's VirutalEntityService, you can create a BaseService
with a specific Quick entity.  That entity is reused for all queries on the
instantiated service. This makes it perfect for handlers and singletons.
This is a basic implementation that forwards calls on the a reset
version of the contained entity.  There may be edge cases that aren't
accounted for.