Check for equivalence of strings changed to compare()
When checking for equality of strings, the "==" or "eq" operator doesn't distinguish between case. As a result, "A" == "a" would return "true".

Therefore the "==" test has been changed to compare(...) == 0 which does consider the case of the strings. As a result, compare( "A", "a") would return "-1", which means the two strings are not identical, even in case. Where as compare( "A", "A") would return "0".