fix(BaseEntity): Return null instead of unloaded entities.
The idea to return unloaded entities was an effort to avoid some of
CFML's null handling.  It has caused more issues, however,
so now we will return null when an entity is not found, either from
a query or from a relationship.

BREAKING CHANGE: `null` is now returned instead of an unloaded
entity when a query returns empty.  You will need to update any
appropriate `isLoaded` checks with `isNull` instead.