feat(Validation): Remove cbvalidation by default
The Quick hooks in to cbvalidation are not as flexible as wanted.
We could either spend the time exposing more of cbvalidation or
we could remove it and leave that up to userland.  Removing it also
opens up the possibility of using a different validation library
if desired.

BREAKING CHANGE: Automatic validation removed. Users desiring the old
behavior can create their own base entity that adds the validation back
in using entity lifecycle methods.