feat(EagerLoading): Add constraints on an eager loaded relationship
Constraints can be added to an eager loaded relationship using a special
struct syntax instead of a string. The struct has the name of the
relationship as the key and a function as the value. The function
will be called with the relationship and can be constrained using
the normal Quick methods. Nested eager loaded relationships can
be specified using the `with` syntax in the constraint callback.