fix(BaseEntity): Account for null values with custom setters
Using `accessors="true"` Quick tries to call all setters seeing
them all as custom setters.  This caused an issue when
retrieving an entity since null values come back as empty strings
which might not match the type of the property and setter.
This change skips the setter if the value is determined to be
a null value for the column (`nullValue=""` on the property).