feat(Scopes): Scopes with an OR combinator are automatically grouped
If a scope has a where clause with an OR combinator it will automatically be grouped in parenthesis.  Scopes without a where clause with an OR combinator will be added to the existing where clauses without grouping as is currently done.

BREAKING CHANGE: Scopes with an OR combinator are now automatically grouped in parenthesis.  If you were relying on the previous behavior, your code will need to be updated.