perf: Increase entity creation speed by at least 2x
Lots of changes, but the biggest was only creating a QuickBuilder
when one was needed.  For instance, when fetching entities from
the database you don't need each of those entities to come with a
QuickBuilder.  It is _highly_ unlikely that those entities will each
need a builder instance if any of them at all.  Now creating a QuickBuilder
creates a new, separate CFC that drastically cuts down on processing time.

BREAKING CHANGE: As a result of splitting QuickBuilder from the BaseEntity,
`addSubselect` calls must be called off of the QuickBuilder instance. This
means in your scopes where `addSubselect()` used to work you now need to
call `qb.addSubselect()` where `qb` is the instance passed into the scope
as the first argument.