fix(QuickBuilder): Fixed accessing QuickBuilder from entities and relationships
Fixed an issue where the old style of chaining constraints off the entity did not work.  The QuickBuilder will now be cached in the entity as well as returning itself.
Fixed a regression where `retrieveQuery` did not return the correct instance from a relationship.
Fixed a bug where calling `.with` returned a QuickBuilder instead of the entity.