fix(Relationships): Return default entities for non-matched eager loaded entites
When eager loading a query, a default entity relationship should be respected
and returned if no match is found in the database.

BREAKING CHANGE: Relationships with a default entity but no match used to
return `null` (or an empty struct when in `asQuery` mode). It will now return
the default entity attributes or memento. (Relationships without default
entities have no change.)