fix(HasManyDeep): Fix for deep entities with compound keys
The issue stemmed from using a nested array as a polymorphic
relationship syntax.  This has been swapped to a struct so that
compound keys can continue to be represented as nested arrays.
Fixes #232

BREAKING CHANGE: Polymorphic relationships used inside `hasManyDeep`
relationships need to be constructed as a `struct`.  Using the
`HasManyDeepBuilder` is a preferred option and requires no code changes.